refactor(#5031): move registry types and restore tests (W5b)
