+++ ARCHITECTURE.md
@@ -1164,6 +1164,7 @@ src/
 │   │   ├── tmux_session_files.rs
 │   │   ├── tmux_watcher.rs
 │   │   ├── tmux_watcher_registry.rs
+│   │   ├── tmux_watcher_registry_restore_tests.rs
 │   │   ├── tui_direct_pending_start.rs
 │   │   ├── tui_prompt_relay.rs
 │   │   ├── tui_task_card.rs
