+++ scripts/audit_maintainability_giant_baseline.toml
@@ -288,7 +288,7 @@
 # carry-forward parameter (`None` here; queued kickoff carries the voice payload via
 # the accepted-replay store reinsert above) plus its one-line rationale comment.
 # Bugfix-only (#3905 direct-dispatch carry-forward).
-"src/services/discord/mod.rs" = 2824 # 2026-08-02 #5031 W5a: 3120 -> 2824 (-296), move-only 분해 (TmuxWatcherRegistry impl 추출; giant_production_loc() 실측). 2026-08-02 #5031 W4b: 3376 -> 3120 (-256), move-only 분해 (lease types/constants/heartbeat + lease clock; 260 production LoC moved and 4-line facade re-export; giant_production_loc() 실측). 2026-08-02 #5031 W4a: 3647 -> 3376 (-271), move-only 분해 (DeliveryLeaseCell impl 추출; giant_production_loc() 실측). 2026-08-01 #5031 W3 r2: 3645 -> 3647 (+2 prod LoC), parent imports only the production cleanup entry point plus the one test-referenced marker; the selector test remains in mod.rs because the coverage gate forbids the path rename (measured test=1651, excluded). 2026-08-01 #5031 W3: 3787 -> 3645 (-142), move-only 분해 (session_idle_cleanup 추출). 2026-08-01 #5031 W2: 4058 -> 3787 (-271), move-only 분해 (skills_scan 추출) + 리뷰 P2 import 정규화(고아 use 제거). 2026-08-01 #5031 W1: 4169 -> 4058 (-111), move-only 분해 (role_map_enrichment 추출). 2026-07-17 #4486: 4168 -> 4169 (+1 prod LoC), reviewable admission — `pub(crate) mod bot_role;` registration for the extracted UtilityBotRole enum (same shape as the prior #3983 session_banner module registration); a module declaration cannot be net-zeroed without deleting unrelated content. Prior: 2026-07-05 #4107: 4161 -> 4168 (+7 prod LoC), reviewable admission — mailbox_try_start_turn wrappers route through persistence-aware active-source purge feedback. Prior: 2026-07-04 #4068: 4159 -> 4161 (+2 prod LoC), reviewable admission — mailbox_snapshot call-site adaptation (peek + default fallback) for the non-creating snapshot fix; no logic beyond the fix. Prior: 2026-07-02 #3805 P2 PR-A: 4158 -> 4159 (+1 prod LoC), reviewable admission — the `make_shared_data_for_tests` fixture literal gains one `two_message_panel_enabled: false` line for the new additive scaffolding flag (no-op, no reader); no logic added to this giant. Prior: #3983 item4 4156 -> 4158 (session_banner module registration).
+"src/services/discord/mod.rs" = 2739 # 2026-08-02 #5031 W5b: 2824 -> 2739 (-85), move-only 분해 (registry 타입군 합류 + restore_tests 경로보존 이관; giant_production_loc() 실측). 2026-08-02 #5031 W5a: 3120 -> 2824 (-296), move-only 분해 (TmuxWatcherRegistry impl 추출; giant_production_loc() 실측). 2026-08-02 #5031 W4b: 3376 -> 3120 (-256), move-only 분해 (lease types/constants/heartbeat + lease clock; 260 production LoC moved and 4-line facade re-export; giant_production_loc() 실측). 2026-08-02 #5031 W4a: 3647 -> 3376 (-271), move-only 분해 (DeliveryLeaseCell impl 추출; giant_production_loc() 실측). 2026-08-01 #5031 W3 r2: 3645 -> 3647 (+2 prod LoC), parent imports only the production cleanup entry point plus the one test-referenced marker; the selector test remains in mod.rs because the coverage gate forbids the path rename (measured test=1651, excluded). 2026-08-01 #5031 W3: 3787 -> 3645 (-142), move-only 분해 (session_idle_cleanup 추출). 2026-08-01 #5031 W2: 4058 -> 3787 (-271), move-only 분해 (skills_scan 추출) + 리뷰 P2 import 정규화(고아 use 제거). 2026-08-01 #5031 W1: 4169 -> 4058 (-111), move-only 분해 (role_map_enrichment 추출). 2026-07-17 #4486: 4168 -> 4169 (+1 prod LoC), reviewable admission — `pub(crate) mod bot_role;` registration for the extracted UtilityBotRole enum (same shape as the prior #3983 session_banner module registration); a module declaration cannot be net-zeroed without deleting unrelated content. Prior: 2026-07-05 #4107: 4161 -> 4168 (+7 prod LoC), reviewable admission — mailbox_try_start_turn wrappers route through persistence-aware active-source purge feedback. Prior: 2026-07-04 #4068: 4159 -> 4161 (+2 prod LoC), reviewable admission — mailbox_snapshot call-site adaptation (peek + default fallback) for the non-creating snapshot fix; no logic beyond the fix. Prior: 2026-07-02 #3805 P2 PR-A: 4158 -> 4159 (+1 prod LoC), reviewable admission — the `make_shared_data_for_tests` fixture literal gains one `two_message_panel_enabled: false` line for the new additive scaffolding flag (no-op, no reader); no logic added to this giant. Prior: #3983 item4 4156 -> 4158 (session_banner module registration).
 # #3305: the local-only pass-through lifecycle-skip (kind hoist + note-only return
 # branch + idle-loop skip) is offset by comment dedup in the same root, lowering the
 # frozen baseline 5438 -> 5429 (-9). Locks in the shrink win (#3028 ratchet).
